+++ src/rxcore/index.ts
@@ -1157,4 +1157,14 @@ export class RXCore {
     public static onGuiScaleListLoadComplete(callback: Function): void {
         RxCore.GUI_scaleListLoadComplete.connect(callback);
     }
+
+    public static setMarkupfromJSON(markupJSON: any): void {
+        RxCore.setMarkupfromJSON(markupJSON);
+    }
+
+    public static setUniqueMarkupfromJSON(markupJSON: any, DocRef : any): void {
+        RxCore.setUniqueMarkupfromJSON(markupJSON, DocRef);
+    }
+
+
 }