+++ src/assets/scripts/rxconfig.js
@@ -5,6 +5,7 @@ var RxConfig = (function() {
     var gui360URL = "rxweb/";
 
     var baseURL = "http://viewserver.rasterex.com/";
+    var apiBaseURL = "http://viewserver.rasterex.com:8080/";
 
     var baseURLBin = baseURL + "RxBinWeb/";
     var baseURLWeb = baseURL + "rxweb/";
@@ -70,6 +71,7 @@ var RxConfig = (function() {
     var getInitial = baseURLBin + "RxCSISAPI.dll?WebClientGetInitial";
 
     return {
+        apiBaseURL: apiBaseURL,
         xmlurl: xmlurl,
 	      xmlurldirect : xmlurldirect,
         xmlurlmarkup : xmlurlmarkup,